+++ index.html
@@ -3,6 +3,9 @@
 <head>
   <script src="js/functions.js"></script>
   <script src="js/main.js"></script>
+  <script src="js/constants.js"></script>
+  <script src="js/util.js"></script>
+  <script src="js/data.js"></script>
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width,initial-scale=1">
   <link rel="stylesheet" href="css/normalize.css">
