+++ js/constants.js
@@ -1,0 +1,34 @@
+const NAMES = [
+  'Антон',
+  'Андрей',
+  'Дима',
+  'Аркадий',
+  'Эля',
+  'Настя',
+  'Маша',
+  'Катя',
+  'Анна',
+  'Миша',
+  'Ксюша',
+  'Мирра',
+  'Ева'
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const DESCRIPTIONS = [
+  'Сижу, ем, никого не трогаю',
+  'Найди кота',
+  'Хочу спать...'
+];
+
+const DESCRIPTION_COUNT = 25;
+
+export { NAMES, MESSAGES, DESCRIPTIONS, DESCRIPTION_COUNT };
